requirements.txt
transformers==3.1.0
wordninja==2.0.0
nltk==3.5
tqdm==4.48.2
spacy==2.3.2
numpy==1.18.5
ujson==1.35
torch==1.6.0
absl_py==0.8.1
absl==0.0
PyYAML==5.3.1
